src/repairs_api/pyproject.toml
[project]
name = "service-desk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cosmos>=4.14.3",
    "fastapi[standard]>=0.123.10",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0",
]
